Ending State Management
