+++ src/index.css
@@ -8,5 +8,5 @@ body {
   -moz-osx-font-smoothing: grayscale;
   background: linear-gradient( 
     0deg
-    , #370617 2% ,#000000 );
+    , #370617 -47% ,#000000 );
 }
